# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CaesarRequestMode = typeof CaesarRequestMode[keyof typeof CaesarRequestMode];


export const CaesarRequestMode = {
  encrypt: 'encrypt',
  decrypt: 'decrypt',
} as const;

export interface CaesarRequest {
  text: string;
  /**
     * @minimum 1
     * @maximum 25
     */
  shift: number;
  mode: CaesarRequestMode;
}

export type VigenereRequestMode = typeof VigenereRequestMode[keyof typeof VigenereRequestMode];


export const VigenereRequestMode = {
  encrypt: 'encrypt',
  decrypt: 'decrypt',
} as const;

export interface VigenereRequest {
  text: string;
  key: string;
  mode: VigenereRequestMode;
}

export type Base64RequestMode = typeof Base64RequestMode[keyof typeof Base64RequestMode];


export const Base64RequestMode = {
  encode: 'encode',
  decode: 'decode',
} as const;

export interface Base64Request {
  text: string;
  mode: Base64RequestMode;
}

export interface SimpleTextRequest {
  text: string;
}

export interface CipherResult {
  input: string;
  output: string;
  algorithm: string;
  mode: string;
  timestamp: string;
}

export interface FrequencyEntry {
  char: string;
  count: number;
  percentage: number;
}

export interface FrequencyAnalysisResult {
  text: string;
  totalChars: number;
  alphabeticChars: number;
  frequencies: FrequencyEntry[];
  mostCommon: string;
  suggestedShift: number;
}

export interface CipherHistoryItem {
  id: string;
  algorithm: string;
  mode: string;
  input: string;
  output: string;
  timestamp: string;
}

export interface CipherHistoryResponse {
  items: CipherHistoryItem[];
  total: number;
}

export type CipherStatsByAlgorithm = {[key: string]: number};

export type CipherStatsByMode = {[key: string]: number};

export interface CipherStats {
  totalOperations: number;
  byAlgorithm: CipherStatsByAlgorithm;
  byMode: CipherStatsByMode;
  mostUsedAlgorithm: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}