# lib/api-zod/src/generated/types/base64Request.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */
import type { Base64RequestMode } from './base64RequestMode';

export interface Base64Request {
  text: string;
  mode: Base64RequestMode;
}