# lib/api-zod/src/generated/types/caesarRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */
import type { CaesarRequestMode } from './caesarRequestMode';

export interface CaesarRequest {
  text: string;
  /**
     * @minimum 1
     * @maximum 25
     */
  shift: number;
  mode: CaesarRequestMode;
}