# lib/api-zod/src/generated/types/cipherHistoryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */

export interface CipherHistoryItem {
  id: string;
  algorithm: string;
  mode: string;
  input: string;
  output: string;
  timestamp: string;
}