# lib/api-zod/src/generated/types/cipherStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */
import type { CipherStatsByAlgorithm } from './cipherStatsByAlgorithm';
import type { CipherStatsByMode } from './cipherStatsByMode';

export interface CipherStats {
  totalOperations: number;
  byAlgorithm: CipherStatsByAlgorithm;
  byMode: CipherStatsByMode;
  mostUsedAlgorithm: string;
}