# lib/api-zod/src/generated/types/frequencyAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */
import type { FrequencyEntry } from './frequencyEntry';

export interface FrequencyAnalysisResult {
  text: string;
  totalChars: number;
  alphabeticChars: number;
  frequencies: FrequencyEntry[];
  mostCommon: string;
  suggestedShift: number;
}