# lib/api-zod/src/generated/types/frequencyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */

export interface FrequencyEntry {
  char: string;
  count: number;
  percentage: number;
}