# lib/api-zod/src/generated/types/simpleTextRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */

export interface SimpleTextRequest {
  text: string;
}