# lib/api-zod/src/generated/types/vigenereRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WiZzryptic API - Professional Cryptography Toolkit
 * OpenAPI spec version: 0.1.0
 */
import type { VigenereRequestMode } from './vigenereRequestMode';

export interface VigenereRequest {
  text: string;
  key: string;
  mode: VigenereRequestMode;
}